% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRMCbinary-package.R
\docType{package}
\name{MRMCbinary-package}
\alias{MRMCbinary-package}
\title{MRMCbinary: Evaluating Diagnostic Accuracy of Binary Medical Tests in Multi-reader Multi-case Study}
\description{
R package \bold{MRMCbinary} is a package aimed at analyzing the performance of diagnostic tests for binary outcome.
}
\references{
Lee, S., Jang, S., and Lee, W. Evaluating Diagnostic Accuracy of Binary Medical Tests in Multi-reader Multi-case Study. \emph{submitted}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/seungjae2525/MRMCbinary}
  \item Report bugs at \url{https://github.com/seungjae2525/MRMCbinary/issues}
}

}
\author{
Seungjae Lee \email{seungjae2525@gmail.com} and Woojoo Lee \email{lwj221@gmail.com}
}
\keyword{package}
