% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRMCbinary.R
\name{MRMCbinary}
\alias{MRMCbinary}
\title{Sensitivity range of the difference in adjusted RMST}
\usage{
MRMCbinary(
  data,
  Reader,
  Modality,
  Case,
  D,
  Y,
  effect,
  interaction = NULL,
  reference.Modality,
  reference.Reader
)
}
\arguments{
\item{data}{A data frame in which contains the reader identifiers (Reader), modality identifiers (Modality), case identifiers (Case), true disease status (D), and binary diagnostic test result (Y).}

\item{Reader}{Variable of reader identifiers.}

\item{Modality}{Variable of modality identifiers.}

\item{Case}{Variable of case identifiers.}

\item{D}{Variable of true disease status.}

\item{Y}{Variable of binary diagnostic test result.}

\item{effect}{Effect one wants to evaluate (one of "Modality", "Reader", and "Both").}

\item{interaction}{If one want to evaluate the interaction effect between modality and reader, interaction = "TRUE", otherwise "FALSE". Specify only when effect is "both". Default: NULL.}

\item{reference.Modality}{Reference in Variable of modality identifiers.}

\item{reference.Reader}{Reference in Variable of reader identifiers.}
}
\value{
An object of class \code{MRMCbinary}. The object is a data.frame with the following components:
\item{N}{Total number of subjects}
\item{N.exposed}{The number of subjects in exposed group}
The results for the \code{MRMCbinary} are printed with the XXXXX function.
To generate the plot of results for the \code{MRMCbinary}, use the XXXXX function.
}
\description{
\code{MRMCbinary()} is the main function of \code{MRMCbinary} package and
performs the analysis the performance of diagnostic tests for binary outcome.
}
\details{
There are four possible methods for our sensitivity analysis.

In general settings,
\itemize{
\item methods="Approx": In general survival analysis setting, if the censoring rate is less than 0.7,
the approximate optimization method can be recommended
because it is much faster than and very accurate as the direct optimization method.
\item methods="Optim": If the censoring rate is greater than 0.7, the direct optimization method can be used as an alternative
because it is implemented as fast as the approximate optimization method.
}

In special settings, some analytic results can be obtained.
\itemize{
\item methods="LP1": When a closed cohort where all subjects are followed up from the same entry time and
only administrative censoring is allowed at the end of follow-up is considered,
high-dimensional optimization problems can be expressed as the well-known linear programming problems,
and thus one can use the analytic solutions for computing the sensitivity range.
\item methods="LP2": Similarly, when the minimum value of censoring times in each group is longer than or equal to tau,
the optimization problems are also transformed to well-known linear programming problems,
and thus one can use the analytic solutions for computing the sensitivity range.
}

See Lee et al. (2025) for details.
}
\examples{
## Load example data
data(VanDyke)

## Return the First Parts of an Object
head(VanDyke)

## Create binary test results (Y_ijk)
VanDyke$Y <- as.numeric(VanDyke$rating >= 3)

}
\references{
Lee, S., Jang, S., and Lee, W. Evaluating Diagnostic Accuracy of Binary Medical Tests in Multi-reader Multi-case Study. submitted.
}
\seealso{
\code{\link[MRMCbinary]{print.MRMCbinary}}, \code{\link[MRMCbinary]{summary.MRMCbinary}}
}
\keyword{methods}
