% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.MRMCbinary.R
\name{print.MRMCbinary}
\alias{print.MRMCbinary}
\title{Print for \code{MRMCbinary} objects}
\usage{
\method{print}{MRMCbinary}(x, ...)
}
\arguments{
\item{x}{An object for class \code{MRMCbinary}.}

\item{...}{Further arguments (currently not used).}
}
\description{
Print the results for object of class \code{MRMCbinary}.
}
\details{
Print the results for object (\code{MRMCbinary} of class \code{MRMCbinary}.
}
\examples{
## Load example data
data(VanDyke)

## Return the first parts of an object
head(VanDyke)

## Extract Unique readers
unique(VanDyke$reader)

## Extract unique modalities
unique(VanDyke$treatment)

## Create binary test results (Y_ijk)
VanDyke$Y <- as.numeric(VanDyke$rating >= 3)


}
\seealso{
\code{\link[MRMCbinary]{MRMCbinary}}, \code{\link[MRMCbinary]{summary.MRMCbinary}}, \code{\link[base]{print}}
}
\keyword{print}
