% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.MRMCbinary.R
\name{summary.MRMCbinary}
\alias{summary.MRMCbinary}
\title{Summary for \code{MRMCbinary} objects}
\usage{
\method{summary}{MRMCbinary}(object, digits = max(1L, getOption("digits") - 3L), ...)
}
\arguments{
\item{object}{An object for class \code{MRMCbinary}.}

\item{digits}{Summary digits. Default: max(1L, getOption("digits") - 3L).}

\item{...}{Further arguments (currently not used).}
}
\description{
Summary the results for object of class \code{MRMCbinary}.
}
\details{
Summary the results for object (\code{MRMCbinary} of class \code{MRMCbinary}.
}
\examples{
## Load example data
data(VanDyke)

## Return the First Parts of an Object
head(VanDyke)

## Create binary test results (Y_ijk)
VanDyke$Y <- as.numeric(VanDyke$rating >= 3)

}
\seealso{
\code{\link[MRMCbinary]{MRMCbinary}}, \code{\link[MRMCbinary]{print.MRMCbinary}}, \code{\link[base]{summary}}
}
\keyword{summary}
