% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRMCbinary-package.R
\docType{package}
\name{MRMCbinary-package}
\alias{MRMCbinary-package}
\title{MRMCbinary: Multi-reader multi-case analysis of binary diagnostic tests}
\description{
R package \bold{MRMCbinary} is a package aimed at comparing the performance of diagnostic tests (i.e., sensitivity and specificity) for binary outcome in multi-reader multi-case (MRMC) study.
}
\references{
Lee, S., Jang, S., and Lee, W. (2025). Evaluating Diagnostic Accuracy of Binary Medical Tests in Multi-reader Multi-case Study.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/seungjae2525/MRMCbinary}
  \item Report bugs at \url{https://github.com/seungjae2525/MRMCbinary/issues}
}

}
\author{
Seungjae Lee \email{seungjae2525@gmail.com} and Woojoo Lee \email{lwj221@gmail.com}
}
\keyword{package}
