% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRMCbinary.R
\name{MRMCbinary}
\alias{MRMCbinary}
\title{Multi-reader multi-case analysis of binary diagnostic tests}
\usage{
MRMCbinary(
  data,
  Reader,
  Modality,
  Case,
  D,
  Y,
  effect,
  interaction = NULL,
  reference.Modality = NULL,
  reference.Reader = NULL
)
}
\arguments{
\item{data}{A data frame in which contains the reader identifiers (Reader), modality identifiers (Modality), case identifiers (Case), true disease status (D), and binary diagnostic test result (Y).}

\item{Reader}{Variable of reader identifiers.}

\item{Modality}{Variable of modality identifiers.}

\item{Case}{Variable of case identifiers.}

\item{D}{Variable of true disease status. It should be set the value to 1 for cases diseased and to 0 for those non-diseased.}

\item{Y}{Variable of binary diagnostic test result. It should be set the value to 1 for cases diagnosed as positive and to 0 for those diagnosed as negative.}

\item{effect}{Effect one wants to evaluate (one of "Modality", "Reader", and "Both").}

\item{interaction}{If one want to evaluate the interaction effect between modality and reader, interaction = "TRUE", otherwise "FALSE". Specify only when effect is "Both". Default: NULL.}

\item{reference.Modality}{Reference in Variable of modality identifiers.}

\item{reference.Reader}{Reference in Variable of reader identifiers.}
}
\value{
An object of class \code{MRMCbinary}. The object is a data.frame with the following components:
\item{CLR_sen}{Conditional logistic regression results for sensitivity.}
\item{CLR_LRT_sen}{Likelihood ratio test from conditional logistic regression results for sensitivity.}
\item{CLR_Score_sen}{Score test from conditional logistic regression results for sensitivity.}
\item{CLR_Wald_sen}{Wald test from conditional logistic regression results for sensitivity.}
\item{Q_MN_sen}{Cochran's Q test (when the number of modalities is greater than 2) or McNemar's test (when the number of modalities is equal to 2) result for sensitivity. This is only included if (1) effect = "Modality", (2) effect = "Reader", or (3) effect = "Both" and interaction = TRUE.}
\item{CLR_spe}{Conditional logistic regression results for specificity.}
\item{CLR_LRT_spe}{Likelihood ratio test from conditional logistic regression results for specificity.}
\item{CLR_Score_spe}{Score test from conditional logistic regression results for specificity.}
\item{CLR_Wald_spe}{Wald test from conditional logistic regression results for specificity.}
\item{Q_MN_spe}{Cochran's Q test (when the number of modalities is greater than 2) or McNemar's test (when the number of modalities is equal to 2) result for specificity This is only included if (1) effect = "Modality", (2) effect = "Reader", or (3) effect = "Both" and interaction = TRUE.}
\item{formula}{Formula used in conditional logistic regression.}
\item{n.reader}{Total number of readers.}
\item{n.modality}{Total number of modalities.}
\item{effect}{Effect one wants to evaluate. See Details.}
\item{interaction}{If one want to evaluate the interaction effect between modality and reader in conditional logistic regression, interaction = "TRUE", otherwise "FALSE". Specify only when effect is "Both". This is only included if effect = "Both" and interaction = TRUE. See Details.}
\item{reference.Modality}{Reference in variable of modality identifiers.}
\item{reference.Reader}{Reference in variable of reader identifiers.}
The results for the \code{SensSpec} are printed with the \code{\link[MRMCbinary]{print.MRMCbinary}} function.
Also, the results for the \code{SensSpec} are summarized with the \code{\link[MRMCbinary]{summary.MRMCbinary}} function.
}
\description{
\code{MRMCbinary()} is the main function of \code{MRMCbinary} package and
can be used to compare sensitivity and specificity of diagnostic tests for binary outcome in multi-reader multi-case (MRMC) study.
}
\details{
There are three effects that can be evaluated:
\itemize{
\item effect="Modality": This is used when the goal is to exclusively evaluate the effect of multiple modalities.
\item effect="Reader": This is used when the goal is to exclusively evaluate the effect of multiple readers.
\item effect="Both": This is used when the goal is to simultaneously evaluate the effect of multiple modalities and multiple readers.
In this case, "interaction" must be specified.
If one want to evaluate the interaction effect between modality and reader in conditional logistic regression, interaction = "TRUE", otherwise "FALSE".
When interaction effects are taken into account, main effects of modalities or readers cannot be interpreted.
}

See Lee et al. (2025) for details.
}
\examples{
## Load example data
data(VanDyke)

## Return the first parts of an object
head(VanDyke)

## Extract Unique readers
unique(VanDyke$reader)

## Extract unique modalities
unique(VanDyke$treatment)

## Create binary test results (Y_ijk)
VanDyke$Y <- as.numeric(VanDyke$rating >= 3)


}
\references{
Lee, S., Jang, S., and Lee, W. Evaluating Diagnostic Accuracy of Binary Medical Tests in Multi-reader Multi-case Study.
}
\seealso{
\code{\link[MRMCbinary]{print.MRMCbinary}}, \code{\link[MRMCbinary]{summary.MRMCbinary}}
}
\keyword{methods}
