% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.MRMCbinary.R
\name{print.MRMCbinary}
\alias{print.MRMCbinary}
\title{Print for \code{MRMCbinary} objects}
\usage{
\method{print}{MRMCbinary}(x, ...)
}
\arguments{
\item{x}{An object for class \code{MRMCbinary}.}

\item{...}{Further arguments (currently not used).}
}
\description{
Print the results for object of class \code{MRMCbinary}.
}
\details{
Print the results for object (\code{MRMCbinary} of class \code{MRMCbinary}.
}
\examples{
## Load example data
data(VanDyke)

## Return the first parts of an object
head(VanDyke)

## Extract Unique readers
unique(VanDyke$reader)

## Extract unique modalities
unique(VanDyke$treatment)

## Create binary test results (Y_ijk)
VanDyke$Y <- as.numeric(VanDyke$rating >= 3)

## Example usage of MRMCbinary function:
# When comparing the sensitivities and specificities between modalities
modality_result <- MRMCbinary(data = VanDyke, Reader = reader, Modality = treatment,
                              Case = case, D = truth, Y = Y, effect = "Modality",
                              interaction = NULL,
                              reference.Modality = "1", reference.Reader = "1")
print(modality_result)

# When comparing the sensitivities and specificities between readers
reader_result <- MRMCbinary(data = VanDyke, Reader = reader, Modality = treatment,
                            Case = case, D = truth, Y = Y, effect = "Reader",
                            interaction = NULL,
                            reference.Modality = "1", reference.Reader = "1")
print(reader_result)

# When comparing the sensitivities and specificities between modalities and between readers together
#  not considering interaction between modalities and readers
both_result_wo_int <- MRMCbinary(data = VanDyke, Reader = reader, Modality = treatment,
                                 Case = case, D = truth, Y = Y, effect = "Both",
                                 interaction = FALSE,
                                 reference.Modality = "1", reference.Reader = "1")
print(both_result_wo_int)

# When comparing the sensitivities and specificities between modalities and between readers together
#  considering interaction between modalities and readers
both_result_with_int <- MRMCbinary(data = VanDyke, Reader = reader, Modality = treatment,
                                   Case = case, D = truth, Y = Y, effect = "Both",
                                   interaction = TRUE,
                                   reference.Modality = "1", reference.Reader = "1")
print(both_result_with_int)

}
\seealso{
\code{\link[MRMCbinary]{MRMCbinary}}, \code{\link[MRMCbinary]{summary.MRMCbinary}}, \code{\link[base]{print}}
}
\keyword{print}
