% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.SensSpec.R
\name{print.SensSpec}
\alias{print.SensSpec}
\title{Print for \code{MRMCbinary} objects}
\usage{
\method{print}{SensSpec}(x, ...)
}
\arguments{
\item{x}{An object for class \code{MRMCbinary}.}

\item{...}{Further arguments (currently not used).}
}
\description{
Print the results for object of class \code{MRMCbinary}.
}
\details{
Print the results for object (\code{MRMCbinary} of class \code{MRMCbinary}.
}
\examples{
## Load example data
data(VanDyke)

## Return the First Parts of an Object
head(VanDyke)

## Create binary test results (Y_ijk)
VanDyke$Y <- as.numeric(VanDyke$rating >= 3)

}
\seealso{
\code{\link[MRMCbinary]{SensSpec}}, \code{\link[base]{print}}
}
\keyword{print}
