% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.SensSpec.R
\name{print.SensSpec}
\alias{print.SensSpec}
\title{Print for \code{MRMCbinary} objects}
\usage{
\method{print}{SensSpec}(x, ...)
}
\arguments{
\item{x}{An object for class \code{MRMCbinary}.}

\item{...}{Further arguments (currently not used).}
}
\description{
Print the results for object of class \code{MRMCbinary}.
}
\details{
Print the results for object (\code{MRMCbinary} of class \code{MRMCbinary}.
}
\examples{
## Load example data
data(VanDyke)

## Return the first parts of an object
head(VanDyke)

## Extract Unique readers
unique(VanDyke$reader)

## Extract unique modalities
unique(VanDyke$treatment)

## Create binary test results (Y_ijk)
VanDyke$Y <- as.numeric(VanDyke$rating >= 3)

## Example usage of SensSpec function:
senspe_result <- SensSpec(data = VanDyke, Reader = reader,
                          Modality = treatment, Case = case,
                          D = truth, Y = Y, digits = 3)
print(senspe_result)

}
\seealso{
\code{\link[MRMCbinary]{SensSpec}}, \code{\link[base]{print}}
}
\keyword{print}
