% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SensSpec.R
\name{SensSpec}
\alias{SensSpec}
\title{Calculate sensitivity and specificity}
\usage{
SensSpec(
  data,
  Reader,
  Modality,
  Case,
  D,
  Y,
  digits = max(1L, getOption("digits") - 3L)
)
}
\arguments{
\item{data}{A data frame in which contains the reader identifiers (Reader), modality identifiers (Modality), case identifiers (Case), true disease status (D), and binary diagnostic test result (Y).}

\item{Reader}{Variable of reader identifiers.}

\item{Modality}{Variable of modality identifiers.}

\item{Case}{Variable of case identifiers.}

\item{D}{Variable of true disease status.}

\item{Y}{Variable of binary diagnostic test result.}

\item{digits}{Summary digits. Default: max(1L, getOption("digits") - 3L).}
}
\value{
An object of class \code{SensSpec}. The object is a data.frame with the following components:
\item{N}{Total number of subjects}
\item{N.exposed}{The number of subjects in exposed group}
The results for the \code{SensSpec} are printed with the xxxx function.
To generate the plot of results for the \code{SensSpec}, use the xxxx function.
}
\description{
\code{SensSpec()} is the function that calculates overall sensitivity and specificity, modality specific sensitivity and specificity, and modality specific and reader specific sensitivity and specificity.
}
\examples{
## Load example data
data(VanDyke)

## Return the First Parts of an Object
head(VanDyke)

## Create binary test results (Y_ijk)
VanDyke$Y <- as.numeric(VanDyke$rating >= 3)

}
\seealso{
\code{\link[MRMCbinary]{print.SensSpec}}
}
\keyword{methods}
